feat: Improve shared-types usage
