+++ shared-types/index.ts
@@ -1,11 +1,6 @@
 export * from './responses';
 export * from './api';
 
-export enum RESPONSE_STATUS {
-  error = 'error',
-  success = 'success',
-}
-
 export enum CATEGORY_TYPES {
   internal = 'internal',
   custom = 'custom',
@@ -46,17 +41,4 @@ export interface CategoryModel {
 export interface TransactionModel {
   type: TRANSACTION_TYPES;
   tx: unknown;
-}
-
-export enum KEYBOARD_CODES {
-  plus = 43,
-  minus = 45,
-  equal = 61,
-  keyE = 101,
-  enter = 13,
-  escape = 27,
-  arrowLeft = 37,
-  arrowTop = 38,
-  arrowRight = 39,
-  arrowBottom = 40,
 }
