+++ src/api/user.ts
@@ -1,8 +1,8 @@
-import { api } from '@/api'
-import * as types from '@/common/types'
+import { api } from '@/api/_api';
+import * as types from '@/common/types';
 
 export const loadUserData = async (): Promise<types.UserRecord> => {
-  let result = await api.get('/user');
+  const result = await api.get('/user');
 
   return result;
-}
+};
