+++ src/js/const/index.ts
@@ -1,3 +1,3 @@
 export { PAGE_TRANSITIONS } from './page-transitions.const';
 export { TRANSACTION_TYPES } from './transaction.const';
-export { ERROR_CODES } from './error-codes.const';
+export { API_ERROR_CODES } from './error-codes.const';
