+++ src/js/records/models/index.ts
@@ -1,4 +1,3 @@
-export * from './category.record';
 export * from './user-currency.record';
 export * from './transaction.record';
 export * from './currency.record';
