+++ src/pages/account/types/monobank/load-latest-transactions.vue
@@ -8,7 +8,7 @@
 </template>
 
 <script lang="ts">
-import { ERROR_CODES } from 'shared-types';
+import { API_ERROR_CODES } from 'shared-types';
 import {
   defineComponent, computed, ref, watchEffect, PropType,
 } from 'vue';
@@ -72,7 +72,7 @@ export default defineComponent({
           type: NotificationType.success,
         });
       } catch (e) {
-        if (e.data.code === ERROR_CODES.forbidden) {
+        if (e.data.code === API_ERROR_CODES.forbidden) {
           addNotification({
             text: e.data.message,
             type: NotificationType.error,
