+++ src/stores/integrations/crypto/binance/index.ts
@@ -5,7 +5,7 @@ import {
   NormalizedAccountData,
   AccountSettings,
 } from 'shared-types/responses/crypto/binance';
-import { api } from '@/api';
+import { api } from '@/api/_api';
 import { normalizeAccountData } from './response-normalizer';
 
 export const useCryptoBinanceStore = defineStore('crypto-binance', () => {
